Bugfix: Fix-constraint now works.
Implemented a mechanism for setting the root cluster so that cluster-merging
takes place in the correct order. For each cluster, a root variable is 
created, which is used in cluster-merging methods to choose which 
cluster to transform, and which too keep fixed. The root-variable 
of the output cluster is related to root-variables of the input
clusters by adding a logical OR method (OrMethod) to the 
method graph.

Rick